backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
supabase==2.3.0
xgboost==2.0.3
scikit-learn==1.4.0
pandas==2.2.0
numpy==1.26.3
joblib==1.3.2
python-dotenv==1.0.0
pydantic==2.5.3
python-multipart==0.0.6
python-dateutil==2.8.2
